github_etl/requirements.txt
requests
psycopg2-binary
PyYAML
